+++ src/hooks/index.ts
@@ -1,3 +1,4 @@
 export * from './useFetch';
 export * from './useShare';
 export * from './useUserStats';
+export * from './useFilterPersistence';